Fix all warnings - Changed CMake compile flags when compiling C translation units,   disabling C++ specific flags - Suppressed warnings for moc generated code, and tor code - Fixed all other warnings

+++ src/libtego/source/core/IncomingRequestManager.cpp
@@ -59,7 +59,7 @@ IncomingRequestManager::IncomingRequestManager(ContactsManager *c)
         auto serviceIdRaw = hostname.chopped(tego::static_strlen(".onion"));
 
         std::unique_ptr<tego_v3_onion_service_id_t> serviceId;
-        tego_v3_onion_service_id_from_string(tego::out(serviceId), serviceIdRaw.data(), serviceIdRaw.size(), tego::throw_on_error());
+        tego_v3_onion_service_id_from_string(tego::out(serviceId), serviceIdRaw.data(), static_cast<size_t>(serviceIdRaw.size()), tego::throw_on_error());
 
         std::unique_ptr<tego_user_id_t> userId;
         tego_user_id_from_v3_onion_service_id(tego::out(userId), serviceId.get(), tego::throw_on_error());
